Add a `do` command.
This can be used to execute multiple commands as one command. Why? Because
aliases can only map to one command. So this will allow aliases to execute
multiple commands in a fairly trivial way:

    alias add src-dis do ; src; dis

This will let the user say `src-dis` to get both a source code listing and
a disassembly for the current stack frame.